Added track splitting.
Pressing the "Split Track" button splits the currently selected track at the
current frame. The selected track is shortened so that it's end frame is set
to the current frame, and a new track is created starting from the current
frame and continuing to the end frame of the original track.

Note: Currently, splitting is only useful for slicing out chunks of a large
track, either to ignore or to flip orientation. In the next few updates, the
actual tracking data will be moved in to the tracklet so that new tracks can
be created by hand, and existing tracks can be relabled.