Trackfixer done.
After the discussion this Thursday about splitting the duties of the
teamview replacement/btf editor and the Keyframer, this should be the last
commit for trackfixer that adds features, any subsequent ones will be bugfixes.

Trackfixer as of right now displays BTF data in sync with video so that manual
editing can be done. Currently, BTF data can be relabled, ignored, or have its
orientation flipped 180 degrees. A temporally contiguous set of tracked data
points for a single ID is represented as a horizontal bar, called a "tracklet".
A tracklet can be split, relabled, have it's orientation flipped, or ignored,
and then written back out to BTF. Further documentation on how to use
trackfixer will be added to the wiki.