Half-way updated to pcl-1.7 and qt5
It looks like pcl-1.7 has API changes that affect us,
and pcl-1.6 has macro's that do not play nice with
qt5's moc. So we're stuck at pcl-1.6 and qt4.

The committed changes are backwards-compatible with
qt4 (with warnings about the new Qt5 option "widgets")
but should make us slightly less unprepared for the
upgrade when it arrives.